Abstracted inspiration engine card functionality
